[Test] Eager cancelation for ProgramTester (#14126)
<!--- 
Thanks so much for your contribution! If this is your first time
contributing, please ensure that you have read the
[CONTRIBUTING](https://github.com/pulumi/pulumi/blob/master/CONTRIBUTING.md)
documentation.
-->

# Description

<!--- Please include a summary of the change and which issue is fixed.
Please also include relevant motivation and context. -->
This PR changes the behavior of `ProgramTester` in
`pkg/testing/integration` package, to detect test failures more eagerly
and then to skip remaining steps. Specifically it checks whether the
test been marked as failed by a validation function (as defined by
[`t.Failed()`](https://pkg.go.dev/testing#T.Failed)).

The rationale is that, for most integration tests, each step assumes
that the previous step was successful. It is simply noisy to continue
with the subsequent steps. A workaround is to call `t.FailNow()` in the
validation function (or use `require`).

An option is provided `ExpectTestFailure` to continue with steps after
test failure (as is the current behavior).

There exists an option `ExpectFailures` to continue with steps after a
pulumi deployment failure. I took such failures to be orthogonal to test
failure.

## Checklist

- [x] I have run `make tidy` to update any new dependencies
- [x] I have run `make lint` to verify my code passes the lint check
  - [ ] I have formatted my code using `gofumpt`

<!--- Please provide details if the checkbox below is to be left
unchecked. -->
- [ ] I have added tests that prove my fix is effective or that my
feature works
<!--- 
User-facing changes require a CHANGELOG entry.
-->
- [ ] I have run `make changelog` and committed the
`changelog/pending/<file>` documenting my change
<!--
If the change(s) in this PR is a modification of an existing call to the
Pulumi Cloud,
then the service should honor older versions of the CLI where this
change would not exist.
You must then bump the API version in
/pkg/backend/httpstate/client/api.go, as well as add
it to the service.
-->
- [ ] Yes, there are changes in this PR that warrants bumping the Pulumi
Cloud API version
<!-- @Pulumi employees: If yes, you must submit corresponding changes in
the service repo. -->

## Example
Here's a snippet of output in the event that a validation function marks
the test as failed (e.g. using an assertion).
```
    ...
    /Users/eronwright/Pulumi/pulumi/tests/examples/program.go:1742: Performing extra runtime validation.
    /Users/eronwright/Pulumi/pulumi/tests/examples/examples_test.go:44:
                Error Trace:    ...
                Error:          Expected nil, but got ...
                Test:           TestAccMinimal
    /Users/eronwright/Pulumi/pulumi/tests/examples/program.go:1744: Extra runtime validation complete.
    /Users/eronwright/Pulumi/pulumi/tests/examples/program.go:1164: Canceling further steps due to test failure
    /Users/eronwright/Pulumi/pulumi/tests/examples/program.go:1349: Destroying stack
    ...
    /Users/eronwright/Pulumi/pulumi/tests/examples/program.go:1365: Test failed, retaining stack '...'
--- FAIL: TestAccMinimal (5.61s)
```