use v3 of setup-dotnet instead of a specific SHA (#14156)
# Description

The specific SHA seems to be failing to download sometimes today. Try to
update it to the latest v3 version to see if that works better.

This seems likely an issue that's currently going on with the GitHub
infrastructure, but it seems to me that we don't need to pin this
version anyway, and we can get the latest fixes for the action this way.