Use an envvar other than GITHUB_REF for publish scripts
In our release workflow, we explicitly set `GITHUB_REF` to the specified input value (e.g. `v3.88.1`).

However, according to GitHub's documentation, we shouldn't be overriding this variable:

https://docs.github.com/en/actions/learn-github-actions/variables#default-environment-variables

> You can't overwrite the value of the default environment variables named `GITHUB_*` and `RUNNER_*`.

The most recent run of of the release workflow is failing when running the `scripts/publish_npm.sh` script, and the error is the same error one would get if this envvar isn't set.

To fix this, let's use a different envvar name, that isn't reserved.