pass ConfigType through RunRequest
Currently we're only passing the config as strings through the
RunRequest method.  This means that the program handling the RPC
can't know the right types, and has to infer it.  This leads to issues
such as https://github.com/pulumi/pulumi-yaml/issues/453.

Passing the types through we'll be able to parse the strings correctly
on the receiving end.