tighten up parsing of property path to avoid accepting typos
The key "my:root.[1]" used to not be accepted, because the `.` before
the index is invalid.  However through https://github.com/pulumi/pulumi/pull/13814/files#r1320306140
and subsequently https://github.com/pulumi/pulumi/pull/14149, the
parsing was loosened.

Presumably we were never checking this at the right level in the first
place, and we only accidentally refused to parse it.  Fix this by
tightening the conditions when parsing the property path.