Track actual error in StepExecutor (#14247)
<!--- 
Thanks so much for your contribution! If this is your first time
contributing, please ensure that you have read the
[CONTRIBUTING](https://github.com/pulumi/pulumi/blob/master/CONTRIBUTING.md)
documentation.
-->

# Description

<!--- Please include a summary of the change and which issue is fixed.
Please also include relevant motivation and context. -->

This results in much nicer errors being bubbled up to tests. User UX is
not changed by this because these errors are always wrapped into Bails
and so not shown to users.

We should make a similar change for step generator errors.