tighten up parsing of property path to avoid accepting typos (#14257)
The key "my:root.[1]" used to not be accepted, because the `.` before
the index is invalid. However through
https://github.com/pulumi/pulumi/pull/13814/files#r1320306140 and
subsequently https://github.com/pulumi/pulumi/pull/14149, the parsing
was loosened.

Presumably we were never checking this at the right level in the first
place, and we only accidentally refused to parse it. Fix this by
tightening the conditions when parsing the property path.

## Checklist

- [x] I have run `make tidy` to update any new dependencies
- [x] I have run `make lint` to verify my code passes the lint check
  - [ ] I have formatted my code using `gofumpt`

<!--- Please provide details if the checkbox below is to be left
unchecked. -->
- [x] I have added tests that prove my fix is effective or that my
feature works
<!--- 
User-facing changes require a CHANGELOG entry.
-->
- [ ] I have run `make changelog` and committed the
`changelog/pending/<file>` documenting my change
<!--
If the change(s) in this PR is a modification of an existing call to the
Pulumi Cloud,
then the service should honor older versions of the CLI where this
change would not exist.
You must then bump the API version in
/pkg/backend/httpstate/client/api.go, as well as add
it to the service.
-->
- [ ] Yes, there are changes in this PR that warrants bumping the Pulumi
Cloud API version
<!-- @Pulumi employees: If yes, you must submit corresponding changes in
the service repo. -->