Fix issue calling nonexistent `_configure` method on external types (#14318)
We recently added the ability to assign default values for nested types
in generated Python SDKs, when the nested types are passed as dicts. We
do so by calling a new static `_configure` method on the input type, to
assign default values.

The problem is that we're emitting calls to `_configure` for external
types, and such types may not have been upgraded yet to the latest
codegen, so they may not have a static `_configure` method yet.

This change makes it so that `_configure` is only called for external
types that have a `_configure`, via a runtime `hasattr` check.

Fixes #14316