[sdkgen/python] Fix calling `_configure` with an Output value
We recently added the ability to assign default values for nested types in generated Python SDKs, when the nested types are passed as dicts. We do so by calling a new static `_configure` method on the input type, to assign default values.

This wasn't working correctly if the value was an `Output[T]` or `Awaitable[T]`. This change fixes the code to call `_configure` inside an `apply`, if needed.

Since the boilerplate code was getting bigger, moved it into a `configure` helper method in `_utilities.py`.