Adds the ability to fetch Pulumi templates from an API which returns zipfiles
We depend upon the user to provide a URL with a .zip suffix
in order to determine which methodology to use, and assume the API
serves this artifact on a PUT endpoint at the given path.

We also provide a clean passthrough of any query parameters, though
providing those on the command line can be a bit cumbersome.