Support 'key.[index]' in property paths
Fixes https://github.com/pulumi/pulumi/issues/14439.

Ideally we would treat this as an error, but it seems providers are
building property paths of this format so we need to keep supporting it.

It is at least now tested, and clearly linked back to this issue so we
don't inadvertantly remove support for it in the future.

Yet another element of the system that could be tightened up in a V4
release.