Make import error a bit more clear (#14468)
<!--- 
Thanks so much for your contribution! If this is your first time
contributing, please ensure that you have read the
[CONTRIBUTING](https://github.com/pulumi/pulumi/blob/master/CONTRIBUTING.md)
documentation.
-->

# Description

<!--- Please include a summary of the change and which issue is fixed.
Please also include relevant motivation and context. -->

This error was a bit confusing because it just said the Parent had no
name, but what was really missing was an entry for that name in the
nameTable telling us the URN.