def: support secret encryption + decryption (#155)
These changes add support for encrypting and decrypting static secrets
in environment definitions. Encryprtion and decryption is handled as a
pre-evaluation pass. The parser expects secrets to be decrypted prior
to parsing, and will issue errors if it detects encrypted secrets.

Encrypted secrets are represented in the document by calls to `fn::secret`
of the form

```yaml
fn::secret:
  ciphertext: <base64-encoded value>
```

In order to support this representation, `fn::secret` now requires that
its argument is a string literal.